fix: permitir stock decimal en productos a granel y corregir isBulk/códigos vacíos
Crear un producto a granel (ej. huevo por kg) fallaba porque el stock se
validaba como entero. De paso se detectó que isBulk/bulkUnit nunca se
guardaban al crear (sí al editar), y que dejar barcode/internalCode en
blanco producía falsos conflictos de duplicado por comparar "" en vez de
null.

+++ apps/backend/src/modules/products/products.routes.ts
@@ -40,9 +40,9 @@ const createSchema = z.object({
   costPrice: z.number().min(0),
   salePrice: z.number().min(0),
   wholesalePrice: z.number().min(0).optional().nullable(),
-  minStock: z.number().int().min(0).default(0),
-  maxStock: z.number().int().min(0).optional().nullable(),
-  currentStock: z.number().int().min(0).default(0),
+  minStock: z.number().min(0).default(0),
+  maxStock: z.number().min(0).optional().nullable(),
+  currentStock: z.number().min(0).default(0),
   trackExpiry: z.boolean().default(false),
   trackBatch: z.boolean().default(false),
   isBulk: z.boolean().default(false),
@@ -164,7 +164,7 @@ router.get('/:id/movements', async (req: Request, res: Response, next: NextFunct
 router.post('/:id/adjust-stock', authorizeMinRole('WAREHOUSE'), async (req: Request, res: Response, next: NextFunction) => {
   try {
     const { physicalQuantity, reason, notes } = z.object({
-      physicalQuantity: z.number().int().min(0),
+      physicalQuantity: z.number().min(0),
       reason: z.string().min(1),
       notes: z.string().optional(),
     }).parse(req.body);
